+++ src/aks-preview/HISTORY.rst
@@ -11,6 +11,15 @@ To release a new version, please select a new version number (usually plus 1 to 
 
 Pending
 +++++++
+* Update the minimum required cli core version to `2.61.0`.
+
+7.0.0b1
+++++++++
+* [BREAKING CHANGE]: Remove support for `az aks get-os-options` command.
+
+6.0.0b1
+++++++++
+* [BREAKING CHANGE]: Remove support for `az aks mesh` egress gateway commands.
 * Add validation to `az aks create` and `az aks update` while modifying the `--ephemeral-disk-volume-type` and `--ephemeral-disk-nvme-perf-tier` values.
 
 5.0.0b4